mozio/requirements.txt
Django==2.2.1
django-filter==2.1.0
django-language-field==0.0.3
django-money==0.14.4
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-gis==0.14
psycopg2-binary==2.8.2
